+++ pokedex/src/pages/detailsPage/LogoPoke.js
@@ -1,0 +1,27 @@
+// function LogoPoke({ background, hoverBackground }) {
+//     const Logo = styled.div`
+//         background: ${props => `url(${props.background}) no-repeat top center`};
+//         height: 25vh;
+//         width: 25vh;
+//         background: rgb(241, 241, 241);
+    
+//       &:hover {
+//         background: ${props =>
+//           `url(${props.hoverBackground}) no-repeat top center`};
+//         height: 25vh;
+//         width: 25vh;
+//         background: rgb(241, 241, 241);
+        
+//       }
+//     `;
+  
+//     return (
+//       <Logo
+//         background={background}
+//         hoverBackground={hoverBackground}
+//         className="logoContent"
+//       />
+//     );
+//   }
+
+//   export default LogoPoke
